tjr214/single-file: Comment out JWT_SECRET syncing between local and remote
Turns out you dont actually need to use the same JWT_SECRET on both local dev and production. Not sure what led me to believe that in the first place... buuuut I didnt fully remove the code just in case there turns out to be a real reason for it.

PEACE