tjr214/coolify: Work on Coolify API integration
Ok! We can setup the project in Coolify, configure the env vars for the server, and deploy both client and server! Yay! BUUUUUT it only works with PUBLIC GitHub repos. The private endpoints for the Coolify API are giving me errors. Either there is a bug in the code or I am doing something wrong. Will reach out to their dev team and see whats-what.

But for now: One more step closer! YES!

PEACE!