tjr214/coolify: Integrate new API stuff with the .env file generation
`.env.coolify` is now generated with extra template variables to handle the new Coolify API features. We also implement functionality for getting this data from the user.

Additionally, the `run_coolify_healthcheck` function was written. It is a handy-dandy little function for telling us if we are online or now! Yeah!

LETS GO!