tjr214/coolify: Implement `get new private key`
So I started the work to get cool-deploy working with Private github repos that are protected with a Private SSH Key. Was coming along well until I needed to Create a New Key. It took me hours to figure out why the JSON was not validating! Every actual newline needed to be replaced by the literal characters `\n` (so the multi-line string becomes one big line). It is finally working though. OH YEAH!

Lets go!

PEACE