tjr214/coolify: Add auto-redeploy for Public and Private Key repos
Apps now auto-redeploy (just like the built-in functionality for the GitHub Apps)! `cool-deploy` officially completes its task!

I just need to clean a few things up and then I can release it! The only thing not working is Github App deployment -- and for that it seems I need to reach out to the dev team for some help.

Ok so this morning WAS good afterall :)

PEACE!